ANN with more traditional arch
previous ann since test was achitected too much like the dANN. better performance with this model, although the dANN won out for effectiveness. ANN wins out for computational efficiency, but doesn't seem to be effective like the dANN at this scale. The dANN is more of a tradeoff than novel, but a deep thinking machine has its uses.